reorganizing.... about is now a simple QMessage.
